# backend/openapi.yaml
openapi: 3.1.0
info:
  title: AI E-Commerce Backend
  version: 0.1.0
paths:
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: Register
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/verify-otp:
    post:
      tags:
      - auth
      summary: Verify Otp
      operationId: verify_otp_api_v1_auth_verify_otp_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: Refresh
      operationId: refresh_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      summary: Logout
      operationId: logout_api_v1_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/forgot-password:
    post:
      tags:
      - auth
      summary: Forgot Password
      operationId: forgot_password_api_v1_auth_forgot_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/reset-password:
    post:
      tags:
      - auth
      summary: Reset Password
      operationId: reset_password_api_v1_auth_reset_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/seller-requests:
    post:
      tags:
      - buyer
      summary: Create Seller Request
      operationId: create_seller_request_api_v1_seller_requests_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellerRequestIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/reports:
    post:
      tags:
      - buyer
      summary: Create Report
      operationId: create_report_api_v1_reports_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/admin/users:
    get:
      tags:
      - admin
      summary: List Users
      operationId: list_users_api_v1_admin_users_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: role
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/users/{id}:
    patch:
      tags:
      - admin
      summary: Patch User
      operationId: patch_user_api_v1_admin_users__id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchUserIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/seller-requests:
    get:
      tags:
      - admin
      summary: List Seller Requests
      operationId: list_seller_requests_api_v1_admin_seller_requests_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /api/v1/admin/seller-requests/{id}:
    patch:
      tags:
      - admin
      summary: Patch Seller Request
      operationId: patch_seller_request_api_v1_admin_seller_requests__id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchSellerRequestIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/reports:
    get:
      tags:
      - admin
      summary: List Reports
      operationId: list_reports_api_v1_admin_reports_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /api/v1/admin/reports/{id}:
    patch:
      tags:
      - admin
      summary: Patch Report
      operationId: patch_report_api_v1_admin_reports__id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchReportIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/audit-logs:
    get:
      tags:
      - admin
      summary: Audit Logs
      operationId: audit_logs_api_v1_admin_audit_logs_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/admin/logs:
    get:
      tags:
      - admin
      summary: Logs
      operationId: logs_api_v1_admin_logs_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    ForgotPasswordIn:
      properties:
        email:
          type: string
          format: email
          title: Email
      type: object
      required:
      - email
      title: ForgotPasswordIn
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginIn:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginIn
    LogoutIn:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: LogoutIn
...
    PatchReportIn:
      properties:
        status:
          type: string
          title: Status
      type: object
      required:
      - status
      title: PatchReportIn
    PatchSellerRequestIn:
      properties:
        action:
          type: string
          title: Action
        reason:
          type: string
          title: Reason
          default: ''
      type: object
      required:
      - action
      title: PatchSellerRequestIn
    PatchUserIn:
      properties:
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
      type: object
      title: PatchUserIn
...
    RefreshIn:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshIn
    RegisterIn:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
        full_name:
          type: string
          title: Full Name
      type: object
      required:
      - email
      - password
      - full_name
      title: RegisterIn
    ReportIn:
      properties:
        target_type:
          type: string
          title: Target Type
        target_id:
          type: integer
          title: Target Id
        reason:
          type: string
          title: Reason
      type: object
      required:
      - target_type
      - target_id
      - reason
      title: ReportIn
    ResetPasswordIn:
      properties:
        email:
          type: string
          format: email
          title: Email
        otp:
          type: string
          title: Otp
        new_password:
          type: string
          title: New Password
      type: object
      required:
      - email
      - otp
      - new_password
      title: ResetPasswordIn
...
    SellerRequestIn:
      properties:
        shop_name:
          type: string
          title: Shop Name
        description:
          type: string
          title: Description
        contact:
          type: string
          title: Contact
      type: object
      required:
      - shop_name
      - description
      - contact
      title: SellerRequestIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VerifyOtpIn:
      properties:
        email:
          type: string
          format: email
          title: Email
        otp:
          type: string
          title: Otp
      type: object
      required:
      - email
      - otp
      title: VerifyOtpIn
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login